+++ CampusTour/ViewControllers/FeaturedViewController.swift
@@ -1,5 +1,5 @@
 //
-//  TopNavBar.swift
+//  FeaturedViewController.swift
 //  CampusTour
 //
 //  Created by Ji Hwan Seung on 3/18/18.
@@ -8,12 +8,10 @@
 
 import UIKit
 
-class SearchViewController: UIViewController, FilterFunctionsDelegate, PopupFilterProtocol {
-    
+class FeaturedViewController: UIViewController, FilterFunctionsDelegate, PopupFilterProtocol {
     let itemFeedViewController = ItemFeedViewController()
     var filterBar: FilterBar!
     var arButton: UIBarButtonItem!
-    
     let searchManager = ItemFeedSearchManager()
     
     //Replace with data from DataManager
@@ -22,6 +20,7 @@ class SearchViewController: UIViewController, FilterFunctionsDelegate, PopupFilt
     private var filterBarCurrentStatus = FilterBarCurrentStatus(Filter.general.rawValue, Filter.date.rawValue)
     private var blackView: UIView!
     
+    //for popupViewController
     private var isModal = false {
         didSet {
             if !isModal {
@@ -33,7 +32,7 @@ class SearchViewController: UIViewController, FilterFunctionsDelegate, PopupFilt
             }
         }
     }
-
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         
@@ -59,6 +58,25 @@ class SearchViewController: UIViewController, FilterFunctionsDelegate, PopupFilt
         self.present(popupViewController, animated: true, completion: nil)
     }
     
+    //Setup filter & search portion of ViewController
+    func setTopNavBar() {
+        searchManager.delgate = self
+        searchManager.attachTo(navigationItem: navigationItem)
+        searchManager.allData = testEvents as [Any] + testPlaces as [Any]
+        
+        arButton = UIBarButtonItem(title: "AR", style: .plain, target: self, action: #selector(openARMode))
+        navigationItem.setRightBarButton(arButton, animated: false)
+        
+        filterBar = FilterBar()
+        view.addSubview(filterBar)
+        filterBar.snp.makeConstraints { (make) in
+            make.top.equalTo(view.safeAreaLayoutGuide.snp.top) //TODO filter bar shouldn't show at start
+            make.leading.equalToSuperview()
+            make.trailing.equalToSuperview()
+            make.height.equalTo(44)
+        }
+    }
+    
     //Setup Feed portion of ViewController
     func setBottomView() {
         addChildViewController(itemFeedViewController)
@@ -74,27 +92,6 @@ class SearchViewController: UIViewController, FilterFunctionsDelegate, PopupFilt
         
         itemFeedViewController.didMove(toParentViewController: self)
     }
-    
-    //Setup filter & search portion of ViewController
-    func setTopNavBar() {
-        searchManager.delgate = self
-        self.searchManager.attachTo(navigationItem: navigationItem)
-        searchManager.allData = testEvents as [Any] + testPlaces as [Any]
-        
-        arButton = UIBarButtonItem(title: "AR", style: .plain, target: self, action: #selector(openARMode))
-        navigationItem.setRightBarButton(arButton, animated: false)
-        
-        filterBar = FilterBar()
-        filterBar.delegate = self
-        view.addSubview(filterBar)
-        filterBar.snp.makeConstraints { (make) in
-            make.top.equalTo(view.safeAreaLayoutGuide.snp.top)
-            make.leading.equalToSuperview()
-            make.trailing.equalToSuperview()
-            make.height.equalTo(0)
-        }
-    }
-    
     
     func openPopupView(_ data: PopupData) {
         view.endEditing(true)
@@ -148,9 +145,10 @@ class SearchViewController: UIViewController, FilterFunctionsDelegate, PopupFilt
     func updateFilterBar(_ status: FilterBarCurrentStatus) {
         filterBarCurrentStatus = status
     }
+
 }
 
-extension SearchViewController: ItemFeedSearchManagerDelegate {
+extension FeaturedViewController: ItemFeedSearchManagerDelegate {
     func didStartSearchMode() {
         self.navigationItem.setRightBarButton(nil, animated: false)
         
@@ -159,7 +157,7 @@ extension SearchViewController: ItemFeedSearchManagerDelegate {
         view.addSubview(popupViewController.view)
         isModal = false
     }
-
+    
     func didFindSearchResults(results: ItemFeedSpec) {
         if self.searchManager.searchIsActive {
             self.itemFeedViewController.updateItems(newSpec: results)
@@ -167,102 +165,12 @@ extension SearchViewController: ItemFeedSearchManagerDelegate {
     }
     
     func didEndSearchMode() {
+        self.navigationItem.setRightBarButton(arButton, animated: false)
+        self.itemFeedViewController.updateItems(newSpec: ItemFeedSpec.testItemFeedSpec)
+        
         popupViewController.removeFromParentViewController()
         isModal = false
         filterBar.buttons.first?.setTitle(Filter.general.rawValue, for: .normal)
         filterBar.buttons.last?.setTitle(Filter.date.rawValue, for: .normal)
-        
-        self.navigationItem.setRightBarButton(arButton, animated: false)
-        self.itemFeedViewController.updateItems(newSpec: ItemFeedSpec.testItemFeedSpec)
-    }
-}
-
-protocol ItemFeedSearchManagerDelegate: class {
-    func didStartSearchMode()
-    func didFindSearchResults(results: ItemFeedSpec)
-    func didEndSearchMode()
-}
-
-class ItemFeedSearchManager: NSObject, UISearchBarDelegate {
-    weak var delgate: ItemFeedSearchManagerDelegate?
-    
-    var allData = [Any]()
-    var searchBar: UISearchBar
-    private(set) var searchIsActive: Bool = false
-    
-    override init() {
-        searchBar = UISearchBar()
-        super.init()
-        
-        searchBar.placeholder = "Search"
-        searchBar.delegate = self
-        
-        //seems dumb--will have to check on this
-        for s in searchBar.subviews[0].subviews {
-            if s is UITextField {
-                s.layer.borderWidth = 1.0
-                s.layer.borderColor = Colors.shadow.cgColor
-                s.layer.cornerRadius = 10
-            }
-        }
-    }
-    
-    func attachTo(navigationItem: UINavigationItem) {
-        navigationItem.titleView = searchBar
-    }
-    
-    func searchBarCancelButtonClicked(_ searchBar: UISearchBar) {
-        searchBar.resignFirstResponder()
-        searchBar.placeholder = "Search"
-        searchBar.endEditing(true)
-        searchBar.setShowsCancelButton(false, animated: false)
-        
-        self.delgate?.didEndSearchMode()
-    }
-    
-    func searchBarTextDidBeginEditing(_ searchBar: UISearchBar) {
-        //UI elements
-        searchBar.setShowsCancelButton(true, animated: false)
-        
-        self.searchIsActive = true
-        delgate?.didStartSearchMode()
-    }
-    
-    func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
-        //update search results
-        let lowercaseText = searchBar.text?.lowercased() ?? ""
-        
-        let filteredItems: [ItemCellModelInfoConvertible] = self.allData.flatMap { dataElement -> ItemCellModelInfoConvertible? in
-            switch dataElement {
-            case let data as Building where data.name.lowercased().contains(lowercaseText):
-                return data
-            case let data as Event where
-                (data.name + data.description).lowercased().contains(lowercaseText):
-                return data
-            default:
-                return nil
-            }
-        }
-        
-        let itemFeedSpec = ItemFeedSpec(sections: [
-            .items(
-                headerInfo: nil,
-                items: filteredItems)
-            ])
-        
-        self.delgate?.didFindSearchResults(results: itemFeedSpec)
-    }
-    
-    func searchBarSearchButtonClicked(_ searchBar: UISearchBar) {
-        searchBar.resignFirstResponder()
-    }
-
-    func searchBarTextDidEndEditing(_ searchBar: UISearchBar) {
-        print("didEndEditing")
-        searchBar.resignFirstResponder()
-        searchBar.placeholder = "Search"
-        searchBar.endEditing(true)
-        self.searchIsActive = false
-        delgate?.didEndSearchMode()
     }
 }
